Fixes #181 / Added `HTTP 401` and `HTTP 409` for feedback for `/register` and `/login` / Refactoring
