+++ app/templates/register.html
@@ -1,11 +1,11 @@
 {% extends "base.html" %}
 {% block scripts %}
 <script src= "{{ url_for('static', filename='js/register.js') }}"></script>
+<script src= "{{ url_for('static', filename='js/ns/namespace.js') }}"></script>
 {% endblock %}
 {% block content %}
 <h3 id="instructions">Register</h3>
 <form id="registerForm">
-    {{ form.csrf_token() }}
     {{ form.username(placeholder='Username', pattern='^\w+( +\w+)*$') }}
     {{ form.password(placeholder='Password', pattern='^\w+( +\w+)*$') }}
     {{ form.confPass(placeholder='Confirm Password', pattern='^\w+( +\w+)*$') }}
